% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AugmentConceptFiles.R
\name{augmentConceptFiles}
\alias{augmentConceptFiles}
\title{Augment Concept Files}
\usage{
augmentConceptFiles(releaseFolder)
}
\arguments{
\item{releaseFolder}{Folder containing a specific release of a data source}
}
\description{
Augment Concept Files
}
\details{
Adds Data Quality and Temporal details to concept data files
}
