% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AugmentDataQualityFiles.R
\name{augmentDataQualityFiles}
\alias{augmentDataQualityFiles}
\title{Augment Data Quality Files}
\arguments{
\item{sourceFolders}{A vector of folder locations that contain the files
exported from Achilles in the ARES Option format (Achilles::exportAO)
to be included in the network data quality index.}
}
\description{
Augment Data Quality Files
}
\details{
Updates dq-result.json files with information on changes from the previous release
}
