% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildDataQualityHistoryIndex.R
\name{buildDataQualityHistoryIndex}
\alias{buildDataQualityHistoryIndex}
\title{Data Quality Historical Indexing}
\usage{
buildDataQualityHistoryIndex(sourceFolder)
}
\arguments{
\item{sourceFolder}{Location of a data source source folder}
}
\description{
Data Quality Historical Indexing
}
