% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildDataQualityIndex.R
\name{buildDataQualityIndex}
\alias{buildDataQualityIndex}
\title{Build the Data Quality Index}
\usage{
buildDataQualityIndex(sourceFolders, outputFolder)
}
\arguments{
\item{sourceFolders}{A vector of folder locations that contain the dq-result.json files
to be included in the network data quality index.}

\item{outputFolder}{Location of the Ares data folder}
}
\value{
A data frame with a the necessary network data quality data
}
\description{
Build the Data Quality Index
}
\details{
Creates a data quality stratified network index
}
