% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildDataSourceHistoryIndex.R
\name{buildDataSourceHistoryIndex}
\alias{buildDataSourceHistoryIndex}
\title{Data Quality Historical Indexing}
\usage{
buildDataSourceHistoryIndex(sourceFolder)
}
\arguments{
\item{sourceFolder}{Location of a data source folder}
}
\description{
Data Quality Historical Indexing
}
