% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildExportQueryIndex.R
\name{buildExportQueryIndex}
\alias{buildExportQueryIndex}
\title{Export Query Details}
\arguments{
\item{outputFolder}{Path to source folder}
}
\description{
Export Query Details
}
\details{
Exports a query index that contains links to the queries used to generate all data summaries.
}
