% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildNetworkPerformanceIndex.R
\name{buildNetworkPerformanceIndex}
\alias{buildNetworkPerformanceIndex}
\title{Network Performance Index}
\arguments{
\item{sourceFolders}{A vector of data source folders}
}
\value{

}
\description{
Network Performance Index
}
