% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildNetworkUnmappedSourceCodeIndex.R
\name{buildNetworkUnmappedSourceCodeIndex}
\alias{buildNetworkUnmappedSourceCodeIndex}
\title{Build Unmapped Source Code Network Index}
\usage{
buildNetworkUnmappedSourceCodeIndex(sourceFolders, outputFolder)
}
\arguments{
\item{sourceFolders}{A vector of data source folders}

\item{outputFolder}{Location of the Ares data folder}
}
\value{
A data frame with a the network unmapped source code index
}
\description{
Build Unmapped Source Code Network Index
}
\details{
Builds an aggregate index of the unmapped source codes across all source folders.
}
