% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildSourceDataQualityDelta.R
\name{buildSourceDataQualityDelta}
\alias{buildSourceDataQualityDelta}
\title{Build Source Data Quality Delta}
\usage{
buildSourceDataQualityDelta(sourceFolders)
}
\arguments{
\item{sourceFolders}{A vector of folder locations that contain the files
exported from Achilles in the ARES Option format (Achilles::exportAO)
to be included in the network data quality index.}
}
\description{
Build Source Data Quality Delta
}
\details{
Generates data quality delta file that contains information on changes of issues between releases
}
