% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompileCohortDiagnosticsResults.R
\name{compileCohortDiagnosticsResult}
\alias{compileCohortDiagnosticsResult}
\title{compileCohortDiagnosticsResult}
\usage{
compileCohortDiagnosticsResult(sourceFolders)
}
\arguments{
\item{sourceFolders}{A vector of folder locations that contain the files
exported from Achilles in the ARES Option format}
}
\description{
#' A short description of what the function does.
}
\details{
Compile CohortsDiagnostics results into files compatible with Ares
}
