% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeCharacterizationDifference.R
\name{ComputeCharacterizationDifference}
\alias{ComputeCharacterizationDifference}
\alias{computeCharacterizationDifference}
\title{Generate characterization difference results.}
\usage{
computeCharacterizationDifference(sourceFolders)
}
\arguments{
\item{sourceFolder}{Path to source folder}
}
\value{
Table of difference results.
}
\description{
Generate characterization difference results.
}
\details{
Computes characterization difference reports for successive data sources releases of a data source.
}
