% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportToParquet.R
\name{exportReleaseToParquet}
\alias{exportReleaseToParquet}
\title{Export Release to Parquet}
\usage{
exportReleaseToParquet(releaseFolder)
}
\arguments{
\item{releaseFolder}{Release folder containing the files exported from
Achilles in the ARES Option format (Achilles::exportAO)}
}
\description{
Export Release to Parquet
}
\details{
Exports tables from DuckDB database generated by the Achilles package
}
