% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportToParquet.R
\name{exportToParquet}
\alias{exportToParquet}
\title{Export to Parquet}
\arguments{
\item{sourceFolders}{A vector of folder locations that contain the files
exported from Achilles in the ARES Option format (Achilles::exportAO)}
}
\description{
Export to Parquet
}
\details{
Exports tables from DuckDB database generated by the Achilles package
}
