% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetIgnoredReleases.R
\name{getIgnoredReleases}
\alias{getIgnoredReleases}
\title{Get Ignored Releases}
\usage{
getIgnoredReleases(sourceFolder)
}
\arguments{
\item{aresDataRoot}{The location of the source folder.}
}
\value{
A list object of ignored data releases.
}
\description{
Get Ignored Releases
}
\details{
Generates a list of data releases that will be ignored when building for a given source
}
