% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetIgnoredSources.R
\name{getIgnoredSources}
\alias{getIgnoredSources}
\title{Get Ignored Sources}
\usage{
getIgnoredSources(aresDataRoot)
}
\arguments{
\item{aresDataRoot}{The location of the ARES data folder.}
}
\value{
A list object of ignored data sources.
}
\description{
Get Ignored Sources
}
\details{
Generates a list of data sources that will be ignored when building
}
