% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSourceReleaseKey.R
\name{getSourceReleaseKey}
\alias{getSourceReleaseKey}
\title{getSourceReleaseKey}
\usage{
getSourceReleaseKey(connectionDetails, cdmDatabaseSchema)
}
\description{
Returns the source release key for this CDM database for use within ARES
}
