% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildNetworkIndex.R
\name{buildNetworkIndex}
\alias{buildNetworkIndex}
\title{Build Network Index}
\usage{
buildNetworkIndex(sourceFolders, outputFolder)
}
\arguments{
\item{sourceFolders}{A vector of folder locations that contain the files
exported from Achilles in the ARES Option format (Achilles::exportAO)
to be included in the network data quality index.}

\item{outputFolder}{The location of the Ares data folder.}
}
\value{
The JSON object that is written to the ARES data folder (outputFolder parameter).
}
\description{
Build Network Index
}
\details{
Generates a network index containing data source history across releases including performance, and data quality metrics.
}
