% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildNetworkPerformanceIndex.R
\name{buildNetworkPerformanceIndex}
\alias{buildNetworkPerformanceIndex}
\title{Build Network Performance Index}
\arguments{
\item{sourceFolder}{Path to source folder}
}
\value{
Network performance results object.
}
\description{
Build Network Performance Index
}
\details{
Builds an index with network performance results for Achilles
and DQD execution across all source folders.
}
