% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildNetworkUnmappedSourceCodeIndex.R
\name{buildNetworkUnmappedSourceCodeIndex}
\alias{buildNetworkUnmappedSourceCodeIndex}
\title{Network Unmapped Source Code Index}
\usage{
buildNetworkUnmappedSourceCodeIndex(sourceFolders, outputFolder)
}
\arguments{
\item{sourceFolders}{A vector of data source folders}

\item{outputFolder}{Location of the Ares data folder}
}
\value{
A data frame with a the network unmapped source code index
}
\description{
Network Unmapped Source Code Index
}
\details{

}
